gridguard-ai/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
sqlalchemy>=2.0.29
pymysql>=1.1.0
cryptography>=42.0.5
xgboost>=2.0.3
scikit-learn>=1.5.0
shap>=0.46.0
pandas>=2.2.2
numpy>=2.0.0
python-dotenv>=1.0.1
websockets>=12.0
httpx>=0.27.0
